typescript and travisci support
