+++ .travis.yml
@@ -1,0 +1,8 @@
+language: node_js
+sudo: false
+node_js:
+  - 10
+install:
+  - npm install
+script:
+  - npm test
