+++ package.json
@@ -2,7 +2,7 @@
   "name": "backend-coding-test",
   "version": "1.0.0",
   "description": "A coding assessment for Xendit backend interviews",
-  "main": "index.js",
+  "main": "index.ts",
   "directories": {
     "test": "tests"
   },
@@ -11,7 +11,16 @@
     "lint": "eslint ./src",
     "coverage": "nyc npm run unit",
     "test": "npm run lint && npm run coverage",
-    "start": "node index.js"
+    "start": "ts-node-script index.ts"
+  },
+  "mocha": {
+    "require": [
+      "ts-node/register"
+    ],
+    "extension": [
+      "ts",
+      "tsx"
+    ]
   },
   "keywords": [
     "Xendit"
@@ -27,14 +36,24 @@
     "yamljs": "^0.3.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.14",
+    "@types/express": "^4.17.8",
+    "@types/mocha": "^8.0.3",
+    "@types/sinon": "^9.0.8",
+    "@types/sqlite3": "^3.1.6",
+    "@types/supertest": "^2.0.10",
+    "@typescript-eslint/eslint-plugin": "^4.6.0",
+    "@typescript-eslint/parser": "^4.6.0",
     "chai": "^4.2.0",
-    "eslint": "^7.12.0",
+    "eslint": "^7.12.1",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-plugin-import": "^2.22.1",
     "mocha": "^8.2.0",
     "nyc": "^15.1.0",
     "pre-push": "^0.1.1",
     "sinon": "^9.2.0",
-    "supertest": "^4.0.2"
+    "supertest": "^4.0.2",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.5"
   }
 }
