+++ src/schemas.ts
@@ -1,4 +1,6 @@
-module.exports = (db) => {
+import { Database } from 'sqlite3';
+
+export default (db: Database): Database => {
   const createRideTableSchema = `
         CREATE TABLE Rides
         (
