+++ tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "module": "es6",
+  "exclude": [
+    "node_modules"
+  ]
+}
